+++ test/ash_backpex/load_select_test.exs
@@ -1,13 +1,13 @@
 defmodule AshBackpex.LoadSelectTest do
   use ExUnit.Case, async: true
   alias AshBackpex.LoadSelectResolver
-  alias AshBackpex.TestDomain.Item
+  alias AshBackpex.TestDomain.AggregateItem
 
   describe "AshBackpex.LoadSelectResolver" do
     test "resolve_fields_and_loads/2 resolves all field types" do
       fields = [name: %{}, note: %{}, id: %{}, most_viewed: %{}, name_note: %{}, user: %{}]
 
-      {load, select} = LoadSelectResolver.resolve(Item, fields)
+      {load, select} = LoadSelectResolver.resolve(AggregateItem, fields)
 
       assert select === [:name, :note, :id]
       assert load === [:most_viewed, :name_note, :user]
@@ -18,10 +18,10 @@ defmodule AshBackpex.LoadSelectTest do
     fields = [foobar: %{}]
 
     msg =
-      "Unrecognized field. :foobar is not a known attribute, relationship, calculation, or aggregate on AshBackpex.TestDomain.Item"
+      "Unrecognized field. :foobar is not a known attribute, relationship, calculation, or aggregate on AshBackpex.TestDomain.AggregateItem"
 
     assert_raise RuntimeError, msg, fn ->
-      LoadSelectResolver.resolve(Item, fields)
+      LoadSelectResolver.resolve(AggregateItem, fields)
     end
   end
 end