+++ test/support/test_domain.ex
@@ -10,6 +10,7 @@ defmodule AshBackpex.TestDomain do
     resource(AshBackpex.TestDomain.User)
     resource(AshBackpex.TestDomain.Comment)
     resource(AshBackpex.TestDomain.Item)
+    resource(AshBackpex.TestDomain.AggregateItem)
     resource(AshBackpex.TestDomain.ReadOnlyEntry)
     resource AshBackpex.TestDomain.NonDefaultPrimaryKeyName
   end